Rename note_encryption folder into domain
Also rename domain.rs file into zcash_note_encryption_domain.rs
