+++ src/domain/zcash_note_encryption_domain.rs
