+++ src/orchard_flavor.rs
@@ -1,6 +1,6 @@
 //! Defines types and traits for the variations ("flavors") of the Orchard protocol (Vanilla and ZSA).
 
-use crate::{circuit::OrchardCircuit, note_encryption::OrchardDomainCommon};
+use crate::{circuit::OrchardCircuit, domain::OrchardDomainCommon};
 
 /// Represents the "Vanilla" variation ("flavor") of the Orchard protocol.  
 #[derive(Debug, Clone, Default)]
